♻️ refactor : 코드 경로 수정
Resolves: #111
